+++ src/components/Header.jsx
@@ -6,7 +6,6 @@ import ButtonBase from '../components/button/ButtonBase'
 import ZeppelinIcon from './icons/ZeppelinIcon'
 import MenuIcon from './icons/MenuIcon'
 import Dropdown from './dropdown/Dropdown'
-// import Searchbar from './searchbar/Searchbar'
 
 const styles = theme => ({
   root: {
@@ -31,6 +30,15 @@ const styles = theme => ({
     color: theme.colors.gray.white.hex,
     height: '125%',
     width: '100%',
+  },
+  menuContainer: {
+    display: 'flex',
+    alignItems: 'center',
+  },
+  menuText: {
+    display: 'none',
+    color: theme.colors.gray.black.hex,
+    paddingRight: `${ theme.spacing.component.s.rem }rem`,
   },
   [`@media (min-width: ${ theme.breakpoints.xs })`]: {
     container: {
@@ -58,6 +66,9 @@ const styles = theme => ({
     icon: {
       display: 'none',
     },
+    menuText: {
+      display: 'block',
+    },
   },
   [`@media (min-width: ${ theme.breakpoints.l })`]: {
     root: {
@@ -75,7 +86,7 @@ const styles = theme => ({
   },
 })
 
-const Header = ({ handleMenu, handleTheme, classes }) => (
+const Header = ({ handleMenu, theme, handleTheme, classes }) => (
   <header>
     <div className={classes.root}>
       <div className={classes.container}>
@@ -87,13 +98,15 @@ const Header = ({ handleMenu, handleTheme, classes }) => (
           <MenuIcon />
         </ButtonBase>
         <ZeppelinIcon className={classes.icon} />
-        {/* <Searchbar /> */}
         <div />
-        <Dropdown
-          menuItems={['zeppelin', 'cat', 'rental']}
-          menuTitle="Theme"
-          onSelect={handleTheme}
-        />
+        <div className={classes.menuContainer}>
+          <p className={classnames(classes.menuText, 'zep-typo-normal-2')}>Theme:</p>
+          <Dropdown
+            menuItems={['zeppelin', 'cat', 'rental']}
+            selected={theme}
+            onSelect={handleTheme}
+          />
+        </div>
       </div>
     </div>
   </header>
@@ -103,6 +116,7 @@ Header.propTypes = {
   classes: PropTypes.object.isRequired,
   handleMenu: PropTypes.func.isRequired,
   handleTheme: PropTypes.func.isRequired,
+  theme: PropTypes.string.isRequired,
 }
 
 export default withStyles(styles)(Header)