+++ src/components/list/__tests__/List.spec.js
@@ -1,0 +1,10 @@
+import React from 'react'
+import { render } from 'react-testing-library'
+import List from '../List'
+
+describe('List', () => {
+  it('renders correctly', () => {
+    const component = render(<List />)
+    expect(component).toMatchSnapshot()
+  })
+})